More refactoring of gui
