+++ renderer/imgui.ini
@@ -4,7 +4,7 @@ Size=400,400
 Collapsed=0
 
 [Window][Dear ImGui Demo]
-Pos=645,20
+Pos=618,20
 Size=526,680
 Collapsed=0
